Move classNameAndId method to utils directory
